+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.2
+
+* Fixes OOfM error in some cases.
+
+## 0.2.1
+
+* BREAKING: `pdfUri` is replaced with `pdfPath` as now `pdfPath` is capable of taking care both URI path and absolute file path.
+* BREAKING: `pdfsUris` is replaced with `pdfsPaths` as now `pdfsPaths` is capable of taking care both URI paths and absolute file paths.
+* Fixed issue of not being able to use absolute file paths.
+
 ## 0.1.1
 
 * Support for rotating, deleting and reordering pdf pages.